#7: Refactor out basic database code into it's own file
