+++ server.py
@@ -2,6 +2,7 @@ import flask
 import json
 
 import racelogic.resultcalculation as rc
+import racelogic.db as db
 from racelogic.names import NAMES
 
 from flask import Flask, request
@@ -33,7 +34,7 @@ SHORTER_FINAL_NAMES = {
 
 
 def _render_page(active_index, selected_date):
-    db_dates = rc.get_all_dates()
+    db_dates = db.get_all_dates()
     _, (active_tab, active_tab_readable, active_tab_icon) = TABS[active_index]
 
     start_lists = []
@@ -71,7 +72,7 @@ def _render_page(active_index, selected_date):
 
 def _render_individual_result_page(selected_date, results):
     active_index = 1
-    db_dates = rc.get_all_dates()
+    db_dates = db.get_all_dates()
     _, (active_tab, active_tab_readable, active_tab_icon) = TABS[active_index]
 
     laptimes_json = _create_laptimes_json(results)
@@ -104,7 +105,7 @@ def _create_laptimes_json(results):
 
 
 def _is_valid_db_date(date):
-    return date in rc.get_all_dates()
+    return date in db.get_all_dates()
 
 
 def _sort_points(all_points, points_per_race, rcclass):
@@ -124,19 +125,19 @@ def index():
 
 @app.get(f"/{START_LISTS_TAB}")
 def start_lists_default():
-    latest = rc.get_all_dates()[0]
+    latest = db.get_all_dates()[0]
     return flask.redirect(f"/{START_LISTS_TAB}/{latest}")
 
 
 @app.get(f"/{RESULTS_TAB}")
 def results_default():
-    latest = rc.get_all_dates()[0]
+    latest = db.get_all_dates()[0]
     return flask.redirect(f"/{RESULTS_TAB}/{latest}")
 
 
 @app.get(f"/{POINTS_TAB}")
 def points_default():
-    latest = rc.get_all_dates()[0]
+    latest = db.get_all_dates()[0]
     return flask.redirect(f"/{POINTS_TAB}/{latest}")
